ft-#69: Improved the structure of the transaction component.
