+++ src/components/App/router.tsx
@@ -21,8 +21,8 @@ import VerifyEmail from "@pages/Auth/VerifyEmail";
 import ResendEmail from "@pages/Auth/ResendEmail";
 import ForgotLoginPasscode from "@pages/Auth/ForgotLoginPasscode";
 import ResetLoginPasscode from "@pages/Auth/ResetLoginPasscode";
-import Transaction from "@pages/Transaction";
-import Details from "@pages/Details";
+import AllTransactions from "@pages/Transaction";
+import Transaction from "@pages/Transaction/Transaction";
 
 const ROUTER = createBrowserRouter([
   {
@@ -96,7 +96,7 @@ const ROUTER = createBrowserRouter([
         path: CLIENT_ROUTES.allTransactions,
         element: (
           <PageWrapper>
-            <Transaction />
+            <AllTransactions />
           </PageWrapper>
         ),
       },
@@ -104,7 +104,7 @@ const ROUTER = createBrowserRouter([
         path: `${CLIENT_ROUTES.allTransactions}/:type/:id`,
         element: (
           <PageWrapper>
-            <Details />
+            <Transaction />
           </PageWrapper>
         ),
       },
